api-1/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
cachetools
mysql-connector-python
sqlalchemy
